backend_pipelines/process/requirements.txt
google-cloud-pubsub
google-cloud-firestore
functions-framework